services/parser/requirements.txt
Flask==1.0.2
flask-sqlalchemy==2.3.2
flask-migrate==2.2.1
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.14.1
mysqlclient==1.3.12
gunicorn==19.8.1
coverage==4.5.1

beautifulsoup4==4.6.0
requests==2.19.1
redis==2.10.6